requirements.txt
pyserial
openhtf